Adjust start code handling to support error recovery.
Our existing start code recognizer was far too generic: we just need to find a bunch of zeroes followed by a single one, optionally bounded by the distance to the next byte in the case of non-error conditions. We don't need to be able to search for arbitrary bits in the bitstream.