video: Restore the H.263 inversion behavior to Sorenson bitstreams.
Experimentally this appears to fix motion vectors on at least one test video. The "unlimited range" of Sorenson bitstreams refers only to the ability to pull pixels from outside the picture boundaries of the previous reference picture.

This isn't a complete fix for motion vectors but it at least fixes most of them